move Slug from helper to component
