+++ plugins/Admin/src/Controller/AppController.php
@@ -7,4 +7,9 @@ use App\Controller\AppController as BaseController;
 
 class AppController extends BaseController
 {
+    public function beforeFilter(\Cake\Event\EventInterface $event)
+    {
+        parent::beforeFilter($event);
+        $this->loadComponent('Admin.Slug');
+    }
 }
