+++ plugins/Admin/tests/TestCase/Controller/Component/SlugComponentTest.php
@@ -1,0 +1,45 @@
+<?php
+declare(strict_types=1);
+
+namespace Admin\Test\TestCase\Controller\Component;
+
+use Admin\Controller\Component\SlugComponent;
+use Cake\Controller\ComponentRegistry;
+use Cake\TestSuite\TestCase;
+
+/**
+ * Admin\Controller\Component\SlugComponent Test Case
+ */
+class SlugComponentTest extends TestCase
+{
+    /**
+     * Test subject
+     *
+     * @var \Admin\Controller\Component\SlugComponent
+     */
+    protected $Slug;
+
+    /**
+     * setUp method
+     *
+     * @return void
+     */
+    protected function setUp(): void
+    {
+        parent::setUp();
+        $registry = new ComponentRegistry();
+        $this->Slug = new SlugComponent($registry);
+    }
+
+    /**
+     * tearDown method
+     *
+     * @return void
+     */
+    protected function tearDown(): void
+    {
+        unset($this->Slug);
+
+        parent::tearDown();
+    }
+}
